+++ app/models/game.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Game
+  def self.table_name_prefix
+    'game_'
+  end
+end