+++ config/initializers/environment_variables.rb
@@ -1,7 +1,7 @@
-required_variables = %w[
-  DATABASE_URL
-  RAILS_MASTER_KEY
-  APP_HOST
+required_variables = [
+  # DATABASE_URL
+  'RAILS_MASTER_KEY'
+  # APP_HOST
 ]
 
 missing_variables = required_variables.select { |var| ENV[var].nil? }