+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Rails project setup with Ruby 3.4.1 and Rails 8.0.1 
 - Added RuboCop for code style enforcement 
-- Added Kamal 2 deployment configuration
+- Added Kamal 2 deployment configuration => not used
 - Created basic landing page with project title "Uniladder"
-- Added environment variables configuration system 
+- Added environment variables configuration system
+- Added Docker installation instructions for deployment => not used
+- Added core models (User, Game::System, Game::Event) with migrations